+++ includes/acf.php
@@ -23,7 +23,18 @@ use function ElasticPress\Serializers\term_data;
  * @return mixed $value
  */
 function parse_acf_field( $field, $value, $data = array(), $base_prefix = '' ) {
-	switch ( $field['type'] ) {
+	//
+	// NOTE:
+	// If $field is false or doesn't have 'type' key
+	// we skip the switch statement and return the original $value
+	// down below
+	//
+	$field_type = false;
+	if ($field && array_key_exists('type', $field) ) {
+		$field_type = $field['type'];
+	}
+
+	switch ( $field_type ) {
 		// The default `image` metadata is just the attachment id
 		// the image array is much more useful.
 		case 'image':
@@ -75,10 +86,21 @@ function parse_acf_field( $field, $value, $data = array(), $base_prefix = '' ) {
 		case 'flexible_content':
 			if (!getenv('ES_SKIP_ACF_PARSING')) {
 				$data = array();
+
 				foreach ( $value as $index => $m ) {
-					$layout_idx = array_search( $m['acf_fc_layout'], array_column( $field['layouts'], 'name' ) );
-					$layout     = $field['layouts'][ $layout_idx ];
-					$idx        = 0;
+					// keys in $field['layouts'] could be integer or string like "layout_62c81b294f5d7"
+					// so we have to search the layout objects themselves to get the
+					// correctly returned key/index
+					$layout_map = array_map(
+						function($field_layout) {
+							return $field_layout['name'];
+						},
+						$field['layouts']
+					);
+					$layout_idx	= array_search( $m['acf_fc_layout'], $layout_map);
+					$layout		= $field['layouts'][ $layout_idx ];
+					$idx		= 0;
+
 					foreach ( $m as $k => $mod ) {
 						if ( 'acf_fc_layout' === $k ) {
 							continue;
@@ -95,10 +117,15 @@ function parse_acf_field( $field, $value, $data = array(), $base_prefix = '' ) {
 							}
 							$f = acf_get_field( $key );
 						}
-						if ( 'clone' === $f['type'] ) {
+
+						// Note: `acf_get_field` may not find $key,
+						// so check $f for false/null values
+						if ( $f && array_key_exists('type', $f) && 'clone' === $f['type'] ) {
 							$f = $f['sub_fields'][ $idx ];
 						}
-						$data[ $index ][ $k ] = parse_acf_field( $f, $mod, $data, $base_prefix );
+
+						$parsed_value = parse_acf_field( $f, $mod, $data, $base_prefix );
+						$data[ $index ][ $k ] = $parsed_value;
 						$idx++;
 					}
 					$data[ $index ]['acf_fc_layout'] = $m['acf_fc_layout'];